Get rid of appProvisioningProfilePath and AppExtension profile path; they can be found via the bundle identifier in the infoPlist
