+++ plugin/src/main/groovy/org/openbakery/configuration/XcodeConfigTask.groovy
@@ -10,7 +10,9 @@ import org.openbakery.Destination
 import org.openbakery.XcodeBuildPluginExtension
 import org.openbakery.XcodePlugin
 import org.openbakery.XcodeProjectFile
-import org.openbakery.AppExtension;
+import org.openbakery.AppExtension
+
+import java.util.regex.Matcher;
 
 /**
  * User: rene
@@ -83,11 +85,43 @@ class XcodeConfigTask extends AbstractXcodeTask {
 	}
 
 	def setProvisioningProfileIdForExtension(String targetName) {
-		def value = project.xcodebuild.getExtensionProvisioningProfileID(targetName)
+		def value = getExtensionProvisioningProfileID(targetName)
 		if (value) {
 			def key = xcodeProjectFile.getProvisioningProfileKeyForTarget(targetName)
 			plistHelper.setValueForPlist(xcodeProjectFile.projectFile, key, value)
 		}
+	}
+
+
+	String getExtensionProvisioningProfileID(name) {
+		def appExtension = project.xcodebuild.appExtensions.find {  it.name.equalsIgnoreCase(name) }
+		if (null == appExtension) {
+			// Couldn't find an extension with that name
+			throw new IllegalStateException("Couldn't find extension with name: ${name}")
+		}
+		return getProvisioningProfileIDFromInfoPlist(appExtension.infoPlistPath)
+	}
+
+
+	String getAppProvisioningProfileID() {
+		return getProvisioningProfileIDFromInfoPlist(project.xcodebuild.infoPlist)
+	}
+
+	def getProvisioningProfileIDFromInfoPlist(String infoPlistPath) {
+		if (infoPlistPath) {
+			String bundleIdentifier = plistHelper.getValueFromPlist(infoPlistPath, "CFBundleIdentifier")
+			File provisioningProfile = project.xcodebuild.getMobileProvisionFileForIdentifier(bundleIdentifier)
+			if (provisioningProfile?.absolutePath) {
+				def cmd = "security cms -D -i ${provisioningProfile.absolutePath}"
+				def profileText = commandRunner.runWithResult(cmd.tokenize())
+				Matcher matcher = profileText =~ /(?m)<key>UUID<\/key>\s*\n\s*<string>([^<]*)<\/string>/
+				if (matcher.find()) {
+					return matcher[0][1]
+				}
+				throw new IllegalStateException("Couldn't parse out the profile ID for ${profile}")
+			}
+		}
+		return null
 	}
 
 	void createXcode5DeviceList() {
