+++ plugin/src/main/groovy/org/openbakery/packaging/PackageTask.groovy
@@ -95,38 +95,6 @@ class PackageTask extends AbstractDistributeTask {
 
 	}
 
-	File getMobileProvisionFileForIdentifier(String bundleIdentifier) {
-
-		def mobileProvisionFileMap = [:]
-
-		for (File mobileProvisionFile : project.xcodebuild.signing.mobileProvisionFile) {
-			ProvisioningProfileIdReader reader = new ProvisioningProfileIdReader(mobileProvisionFile, project)
-			mobileProvisionFileMap.put(reader.getApplicationIdentifier(), mobileProvisionFile)
-		}
-
-		for ( entry in mobileProvisionFileMap ) {
-			if (entry.key.equalsIgnoreCase(bundleIdentifier) ) {
-				return entry.value
-			}
-		}
-
-		// match wildcard
-		for ( entry in mobileProvisionFileMap ) {
-			if (entry.key.equals("*")) {
-				return entry.value
-			}
-
-			if (entry.key.endsWith("*")) {
-				String key = entry.key[0..-2]
-				if (bundleIdentifier.toLowerCase().startsWith(key)) {
-					return entry.value
-				}
-			}
-		}
-
-		return null
-	}
-
 
 	def addSwiftSupport(File payloadPath,  String applicationBundleName) {
 
@@ -239,7 +207,7 @@ class PackageTask extends AbstractDistributeTask {
 
 		String bundleIdentifier = plistHelper.getValueFromPlist(infoPlist.absolutePath, "CFBundleIdentifier")
 
-		File mobileProvisionFile = getMobileProvisionFileForIdentifier(bundleIdentifier);
+		File mobileProvisionFile = project.xcodebuild.getMobileProvisionFileForIdentifier(bundleIdentifier);
 		if (mobileProvisionFile != null) {
 			File embeddedProvisionFile
